+++ src/app/add-snippet/_components/add-snippet-form.tsx
@@ -18,6 +18,7 @@ import { useForm } from "react-hook-form";
 import * as z from "zod";
 import { addSnippetAction } from "../../_actions/snippet";
 import LanguageDropDown from "./language-dropdown";
+import { catchError } from "@/lib/utils";
 
 const formDataSchema = z.object({
   codeLanguage: z
@@ -89,16 +90,8 @@ export default function AddSnippetForm({ lang }: { lang: string }) {
       });
 
       form.reset();
-    } catch (err: any) {
-      console.log(err);
-      toast({
-        title: "Error!",
-        description: "Something went wrong!" + err.message,
-        duration: 5000,
-        style: {
-          background: "hsl(var(--destructive))",
-        },
-      });
+    } catch (err) {
+      catchError(err);
     }
   }
