+++ src/app/dashboard/components/recentRaces.tsx
@@ -5,7 +5,7 @@ import type { Result } from "@prisma/client";
 
 import { type ColumnDef } from "unstyled-table";
 import Link from "next/link";
-import { DataTable } from "@/components/data-table";
+import { DataTable } from "@/components/data-table/data-table";
 import {
   Tooltip,
   TooltipContent,
