+++ src/app/review/review-buttons.tsx
@@ -35,7 +35,7 @@ export function ReviewButtons({ snippetId }: { snippetId: Snippet["id"] }) {
         onClick={async () => {
           setIsDeleting(true);
           try {
-            await deleteSnippetAction({ id: snippetId });
+            await deleteSnippetAction({ id: snippetId, path: "/review" });
           } catch (err) {
             console.log(err);
           } finally {
